Increase travis thresholds for Mongo 4.0
Mongo 4.0 significantly changes the test runtimes. 2x was too much of a
jump, so this is approximately 60s more than the last successful run
time. That comes out to (test time used to calculate in parentheses):

24% increase for "Unit Tests (Python 2.7 MongoDB 4.0)" (812s)
37% increase for "Integration Tests (Python 2.7)" (903s)
95% increase for "Lint Checks, Packs Tests (Python 2.7)" (490s)
71% increase for "Unit Tests, Pack Tests (Python 3.6)" (1101s)
252% increase for "Integration Tests (Python 3.6)" (1031s)

Note that this PR also fixes the Python 3.6 integration tests to
actually run in python 3.6 instead of 2.7, so that might account for
some of the significant test time increase.

When we move to python 3.7 that will probably decrease the py3 times.