Clean up Makefile task names for easier targetting
The .circle/add-itest-user.sh was actually running for all tasks, so
this removes the if statement.

The permissions workaround needed to run for packs-tests and integration
tests, but there was no easy way to target those unless we separate the
py3 packs-tests from the py3 unit tests. That also makes it much easier
to reason about the different travis environments to understand how they
map to make targets and tox envs.

That also made it clear that one of the tox envs was not getting
pre-built, so this adjusts install-requirements.sh to account for it.