Update make COMMAND_THRESHOLDS
The timing of some make targets is higher than the threshold. For
ci-py3-integration, it looks like that threshold is far too short for
what the tests regularly take. So, for the jobs that are signifcantly
different, this adjust the thresholds to be about 60s more than the last
successful build in travis.

make target: orig threshold=>new threshold (reference duration)

ci-checks ci-packs-tests:   280=>430 (369)
...-py3-unit ci-py3-pac...: 680=>750 (688)
ci-py3-integration:         310=>770 (716)