Fix empty with items list resulting in parent workflow stuck in running
Fix a bug where passing an empty list to a with items task in a subworkflow causes the parent workflow to be stuck in running status. The workflow engine usually skips execution of the with items task because the list is empty. The issue here is that when the task is marked complete, the status change in the task wasn't published which will triggered additional logic to handle completion and move the parent workflow forward. This patch publishes the completeion status of the task now.