Update progress of workflow execution to the output stream
Currently, progress of workflow execution is written to the workflow service and action runner log files. Update the logging in the orquesta runner and workflow service to write the execution progress to the action execution output table which will be available in the streaming API and when tailing execution output.