[BUGFIX] Use pip 20.0.2 to build virtualenvs
We have already been using pip==20.0.2 for testing since st2 v3.2.0.
virtualenv embeds the wheel for pip, so we need to adjust the pinned
version of virtualenv to update which version of pip is used to create
virtualenvs (eg for packs).

This also clarifies the comment about why we're pinning virtualenv.