Update the field and implement another approach which uses additional header for the binary field value.
This header tells us which serialization format and compression (if any)
is used for a specific field value.

Using a header format gives us more, flexibility, makes it more future
proof (e.g. ability to change the format in the future) and also ability
to implement things such as per-field compression.