Implement dict value change tracking for our custom JSONDictField.
This dict value tracking allows us to track when a dict item value has
changed and only write the value to the database on existing document /
model update in case it has changed.

This is a very important property since it allows us to implement
efficient partial document updates.

With that change, JSONDictField now also works in exactly the same
manner as existing mongoengine DictField field type.

Also add tests for various edge cases which would fail if value change
tracking was not correctly implemented or working.