Use sys.stdout.buffer instead of sys.stdout.
sys.tdout.buffer works with bytes where as sys.stdout works with
unicode.

orjson returns bytes and this way we avoid unncessary conversation back
and forth between bytes and unicode.

In addition to that, using bytes means it will also work correctly if
the system locale is not set to utf-8.