Add additional unit tests for  the utf-8 bug.
Keep in mind that API tests are problematic - we don't exercise exactly
the same code paths as we do when running outside tests because the
tests itself utilize webtest module which has the same bug as webob.

This means I needed to patch the code so it utilizes the same version of
the functions as we do in prod.

Not ideal and also shows why it's importat we also have actual end to
end tests for the api.