Update "result_size" field for action execution and live action DB model inside action runner at the end after save.
I was hoping we will be able to avoid one additional serialization, but
sadly we can't if we don't want to massively hack and monkey patch
mongoengine.

And that monkeypatching is not worth it since serialization is fast
enough.

To put things into perspective - takes takes 7ms for 4 MB result which
is nothing compared to other DB operations durations. And for smaller
results it even gets to the sub ms aka nanosecond range.