Add new ?max_result_size query parameter to the GET /v1/executions/<id> API endpoint.
With this query parameter, user can decide to exclude execution.result
field from the response in case it exceeds this size.

This allows us to implement efficient execution retrieval in the client
(st2web) side and avoid additional queries / API operations on the
client side - aka win-win situation (reduce the load on the client and
the server).