Workaround modified GitHub Actions PATH var
We relied on ~/.local/bin being in PATH, which was added in:
https://github.com/actions/virtual-environments/pull/2982

But, then they reverted that because it wasn't cross-shell consistent:
https://github.com/actions/virtual-environments/pull/2982

So, vars were getting loaded from /etc/profile.d, then they moved it
back into ~/.bashrc. But, by default GHA workflows do not source
~/.bashrc.

Who knows how else they are going to mess with the PATH, so just write
our own.