pants: disambiguate test file dependencies
We do large sets of files because maintaining manual per-file mapping of
dependencies would be a lot of pointless work.

Hopefully a future version of pants will have features that make
disambiguation more precise, integrating better with dependency
inferrence. When that happens we can revert this commit in favor of
telling pants how to infer the ambiguous deps instead of hard-coding the
deps for a large set of files that might not need the dep.