Add a workaround for tests getting stuck under Python 3.8 due to pymongo code.
It turns out the issue is related to us not performing monkey patching
early enough in the tests.

This is only an issue for tests, for actual service entry point code,
the code is already structured correctly so we perform monkey patching
as of of the first things we do.