Add unit tests to cover masking of output for various action runners
The output_key for each action runners are different. Python runner writes
output to "result", HTTP runner writes output to "body", and Orquesta
runner writes output to "output". Additional unit tests are added here
to make sure secrets in output are masked appropriately for different
runners. The output_key from the runner model should be used to
dynamically determine which field to apply the masking to.