Refactor mask_secret_output function to be more robust
The current mask_secret_output function in the output_schema util module
is too fragile with the multiple if else on checking runner type to
determine if there is an output schema defined. For example, orquesta
runner has an output_schema and output_key defined but then the if else
already introduced a bug. The function is refactored to evaluate if
output key and output schema is defined and then proceed to masking
secret value if an output parameter is marked as secret in the output
schema. Additional unit test is added to cover orquesta runner. The unit
tests for the mask_secret_output function is also simplified.