Fix masking of secret output for workflows
Add the action.output_schema and runner.output_key to attributes to be
included in the API query otherwise secrets are not masked in the output
of workflow action. Update the output schema util to be more robust in
terms of handling result that is either empty, wrong type, or missing
output key.