Make schema validation optional
There are a lot of issues, so we're only partially validating the spec
now. We still validate with prance, but skip checking x-api-model
because there are so many legacy issues.

I looked at adding the x-api-model ... but wow, we haven't been adding
that for a long time. And there are open issues about it:
https://github.com/StackStorm/st2/issues/3575
https://github.com/StackStorm/st2/issues/3788

So, we just make it possible, but optional, to run the schema validation.