Import PATH/NAME for fixture packs: dummy_pack_1, dummy_pack_10
In order for pants to detect which changes need to be re-run after changing
fixtures, pants needs to be able to infer the relevant dependencies.

We could use a super-level target to depend on everything, but then we lose
the fine-grained invalidation. The only way to tell pants that a test file
is using a particular set of fixtures is via python imports.

So, this adds some boilerplate python in every pack dir:
__init__.py
fixture.py

We cannot use pack.yaml as a source root marker any more or importing these
files fails. We'll have to implement something else for the main packs under
contrib/.