add secret to runner overridable parameters list
Add secret to the list of overridable parameters to avoid the error -  attribute “secret” for the runner parameter “cmd” in action “core.local” cannot be overridden.

Why is this required -

We want to create action core.winrm_ps_cmd_secure which is same as core.winrm_ps_cmd where the input parameter cmd is hidden. We have a use case where we need to supply credentials inside the cmd section so it's not visible in stack storm UI after execution.