update pants-plugins/sample_conf to pants 2.15
The Fmt/Lint plugin API changed in pants 2.15

It now requires a Subsystem with a SkipOption.
I took advantage of this to consolidate the constants and all
the address and pex request definitions in the Subsystem subclass.

It now requires 2 rules instead of 1: to partition, then fmt/lint.
That means the structure of the rule requests had to change
slightly. The tests also had to account for the modified rules.