fix: packaging migration scripts
Running `./pants package ::` results in the following error. We included
the v3.5 migration script in `setup.py` and in the `pants_distribution`.
But, that script does not end in `.py`, so the `python_sources` target
did not "own" it. In order to correct this, we explicitly include the
extension-less binaries.

> ValueError: The explicit dependency
> `st2common/bin/migrations/v3.5/st2-migrate-db-dict-field-values` of
> the target at `st2common:st2common` does not provide enough address
> parameters to identify which parametrization of the dependency target
> should be used.
> Target `st2common/bin/migrations/v3.5:v3.5` can be addressed as:
>   * st2common/bin/migrations/v3.5:v3.5
>   * st2common/bin/migrations/v3.5/__init__.py
>   * st2common/bin/migrations/v3.5/st2_migrate_db_dict_field_values.py

In fixing this, I also noticed that we did not include the v3.8
migration script, so I marked that executable and included it as well.