chore(pants): reclassify st2common/tests/unit/base.py as test_utils not sources
It makes more sense for the files in the tests directory to be "test_utils"
instead of regular sources. It does not make a functional difference to pants,
but I feel it captures the intent of those files better.

The canonical example of "test_utils" is conftest.py which we might switch to
later once we finish transitioning to pytest.