tests: clarify that test_actions_registrar only uses generic fixture pack
The mock enforces that only the generic pack path gets returned for
all of the tests in st2actions/tests/unit/test_actions_registrar.py
even though it called fixtures_loader.get_fixtures_base_path().
Make that even clearer by importing the base path from
st2tests.fixtures.generic.fixture

I tried a variety of ugly hacks to depend on many of the
packs in that directory. It is a relief to not use those
hacks now that I understand that only the generic pack
matters.