Avoid virtualenv/bin/st2-run-pack-tests assumption in CI
If the virtualenv is built by pants, the PEP 660 wheels will not include
any of the scripts (like st2-run-pack-tests). Only entry_point scripts
will be installed in the virtualenv. So, do not assume that a dev venv
will have st2-run-pack-tests. Also, adjust path assumptions to work with
either kind of venv.