Fix `packs.get` action
Addresses #6225. This should allow users to specify a branch when using
the `packs.get` command. This addresses the assumption in the code that
the primary branch of a git repository corresponding to an action is
`master`.

* Add `branch` parameter to `get.yaml`. Default to `master` to maintain
current functionality/expectations.
* Update `run` in `pack_mgmt.get_install.py`. Now accepts `branch` param
and uses an `fstring` to represent the in `git rev list`.