coord.get_members returns NoOpAsyncResult(iterable)
get_members should return a list or tuple, not a string.
I noticed while debugging that get_members output ended up
as ['m', 'e', 'm', 'b', 'e', 'r', '-', '1'] because it
listified the string. So, use a tuple when creating
the mocked NoOpAsyncResult so it is closer to the actual
return values.