Add st2tests.config.coord_opts_as_env_vars for itests
Integration tests need to start subprocesses that use the
same redis as the test. This matters when running under
pantsbuild, because pants runs several instances of pytest
in parallel. This PR prepares to add support, to disambiguate
test runs--similar to the database logic--using the env var:
ST2TESTS_PARALLEL_SLOT.
In any case, when an integration test runs production
code in a subprocess, the production code does not use--
and should not use--any ST2TESTS_* vars, meaning the
subprocess ends up using what is configured in the conf
file instead of the (planned) parallel-safe coordinator.

Thanks to a new-ish oslo_config feature, we can now
update config via env variables. So, make use of that
in integration tests to override the conf-file provided
values with test-provided values.