test: Add ?namespace=... to redis:// uri for parallel testing
This will make it safer to run tests in parallel as the
tests will have separate service registration and coordination locks.
To do that, this adds a namespace= query param to the redis url
that includes the slot number that pants creates to allow tests to run in parallel more safely.
The namespace is used by the tooz library's redis driver as a prefix
for all the redis keys. Thus, multiple tests can use the same
redis database without clobbering each other's keys.