copy webtest pin from lockfiles/st2.lock
This fixes CircleCI.

webtest bumped their dep on waitress to a version that is not available
to all of our interpreters.
https://github.com/Pylons/webtest/commit/090c5b27863b502b91e038ddf2a5649cd5bde2d8

The lockfile has a version locked version that does not have this issue.
The pin was already copied to test-requirements.txt, so this just copies
it to the other requirements files as well. This is not a significant
issue because webtest is only required by st2tests.