Packaging: deb/*.sh Remove logrotate config handling
/etc/logrotate.d/st2 is already recorded in conffiles,
so let dpkg handle conf preservation and restoration.

Besides, there was a bug in the code anyway rendering this logic useless.
post-remove.sh should have had
mv /etc/logrotate.d/st2 /etc/logrotate.d/st2-pkgsaved.disabled
but was missing the first argument, so it did nothing.
This was hidden by the '|| true' that made it always pass.
Since the logic doesn't work, and it is not necessary, just drop it.