Packaging: Remove set_permissions functions
A comment in the pre-install.sh files said:

> Fix directories permissions on upgrade (different across maint scripts!)
> NB! USED FOR COMPATIBILITY ON UPGRADE FROM PREVIOUS VERSIONS OF PACKAGES.
> NB! In future package releases reseting permissions SHOULD BE REMOVED.

This has been there for many releases. So, let's just remove this.
Instead, we rely on the permissions that the st2 package provides,
which are defined in BUILD nfpm_content_* metadata.

This has the benefit of vastly simplifying the deb install scripts
as the upgrade stamp file is no longer required.
So, there is no longer a dependency between pre-install and post-install scripts.