packaging: Use pex_binary(executable=...) instead of pex premable
The preamble method was hacky and had some uncomfortable sharp edges.
The pex cache does not know about the preamble which can break running
the pex to unpack itself in some cases (like if you use PEX_TOOLS to
inspect the pex before unpacking it).

So, switch to a python script that will run when the pex is executed.
This script has access to the full venv, thanks to pex, before we have
even unpacked it. So, we can use deps like oslo.config safely.
However, using a pex entry_point/script/executable causes the script
to be included in the extracted venv. If we import any st2 code, then
that can make pants add other sources or conf files that we do not
want in the pex (other than via the wheels of our sources). So, I left
a note warning against importing st2 code.