Packaging: speed up el8 package build using zstd compression
There was a multi-minute difference between el8 and el9.
The compression algorithm is the only significant difference, so
I timed building the same el8 package with different algorithms.
- xz and lzma were the slowest at over 3 minutes.
- gzip and zstd were the fastest at under 1 minute (at any level).

The size differences were marginal, but zstd:best was the closest
match in size to xz, so I used that to keep the package size as
small as possible while speeding up the package build process.